pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "porypal"
version = "3.0.0"
description = "Palette toolchain for Gen 3 Pokémon ROM hacking — extract, convert, and batch-process sprites for pokeemerald/pokefirered."
readme = "readme.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "prison_lox" }]
keywords = ["pokemon", "romhack", "palette", "gba", "pokeemerald", "sprite"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    # Image processing — pure Python, no Qt dependency in the model
    "Pillow>=11.0",

    # Config
    "PyYAML>=6.0",

    # Palette extraction (v3 pillar 1)
    "numpy>=1.24",
    "scikit-learn>=1.3",

    # CLI (v3 pillar 3)
    "click>=8.1",

    # Web server
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
# GUI — separate so the CLI and model work without a display
gui = [
    "PySide6>=6.6",
]
# Web demo — Gradio for HuggingFace Spaces
web = [
    "gradio>=4.0",
]
# Dev tooling
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",          # linter + formatter, replaces flake8+black
    "pyinstaller>=6.0",   # for building .exe releases
]
# Everything at once
all = ["porypal[gui,web,dev]"]

[project.scripts]
# Installs `porypal` as a CLI command after `pip install porypal`
porypal = "porypal.cli:main"

[project.gui-scripts]
# Installs `porypal-gui` as a desktop launcher (no console window on Windows)
porypal-gui = "porypal.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["porypal*", "controller*", "model*", "view*"]

# ---------- Ruff (linter/formatter) ----------
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]   # pycodestyle + pyflakes + isort
ignore = ["E501"]                # line length handled by formatter

# ---------- Pytest ----------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
